backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
asyncpg
python-jose
passlib[bcrypt]
pydantic-settings
email-validator
python-multipart